requirements.txt
# Flask Backend
flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Logging (optional but recommended)
python-json-logger==2.0.7

# Development Dependencies (optional)
# pytest==7.4.3
# black==23.12.1
# flake8==6.1.0
